fixing fb login attempt 1
