+++ db/index.js
@@ -1,17 +1,14 @@
 const { Sequelize, connection } = require('./connection');
-// import models
-const Users = require('./models/users');
-const { Babeers } = require('./seed');
 
-// associations
+//associations
 
 const sync = () => {
   return new Promise((res, rej) => {
     connection
-      .sync()
+      .sync({force: false})
       .then(() => res(connection))
       .catch(e => rej(e));
   });
 };
 
-module.exports = { connection, sync, Sequelize, Babeers };
+module.exports = { connection, sync};
