+++ index.html
@@ -9,5 +9,6 @@
   <body>
     <div class="container">
       <div id="root"></div>
+    </div>
   </body>
 </html>
