--- server/db/syncForce.js
@@ -1,4 +1,0 @@
-const { connection } = require('.')
-
-// models
-connection.sync({ force: true })
