+++ server/index.js
@@ -1,5 +1,7 @@
 const server = require('./server');
 
+
 const port = process.env.PORT || 3000;
+
 
 server.listen(port, () => console.log(`listening on port ${port}`));
